+++ test/entities/entity.test.ts
@@ -38,6 +38,7 @@ describe('Entity', (): void => {
         expect($entity.getChangeLevel()).toBe(ChangeLevel.Patch);
         expect($entity.getPriority()).toBe(Priority.Low);
         expect($entity.isIgnored()).toBeFalsy();
+        expect($entity.isEscaped()).toBeFalsy();
         expect($entity.isEmpty()).toBeFalsy();
     });
 
@@ -68,5 +69,11 @@ describe('Entity', (): void => {
 
         expect($entity.isIgnored()).toBeTruthy();
         expect($entity.isEmpty()).toBeTruthy();
+    });
+
+    it('Escape', (): void => {
+        $entity.escape();
+
+        expect($entity.isEscaped()).toBeTruthy();
     });
 });